dartsim: Add support for joints in worlds (#501)
As of gazebosim/sdformat#1117, SDFormat supports joints under world (//world/joint in xpath). But in gz-physics, Entities such as joints are only available in a Model. To support world joints, instead of creating a new type of entity, we have chosen to implement a scheme where the world behaves like a model via the WorldModelFeature feature. Given a world, `w`, we can do `model = w->GetWorldModel()` to get the model proxy. We can then use any of the APIs supported by the features available in the physics engine, e.g., `model->GetJoint("j1")` where "j1" is a joint directly under <world> in the SDFormat.

Signed-off-by: Addisu Z. Taddese <addisu@openrobotics.org>
Co-authored-by: Marco A. Gutierrez <marcogg@marcogg.com>